Move test impls to test package
